+++ src/App.tsx
@@ -1,6 +1,6 @@
 import { Flex, useMediaQuery } from "@chakra-ui/react";
 
-import Contents from "./components/PostDetail";
+import PostDetail from "./components/PostDetail";
 import DrawerSideBar from "./components/DrawerSideBar";
 import SideBar from "./components/SideBar";
 
@@ -10,7 +10,7 @@ function App() {
   return (
     <Flex w="100%" h="100vh" bgGradient="linear(to-b, gray.100, purple.100)">
       {desktop ? <SideBar /> : <DrawerSideBar />}
-      <Contents />
+      <PostDetail />
     </Flex>
   );
 }